Change stats recv/send to double, so they can hold truly massive numbers of transferred bytes (they would overflow after about 6 weeks on a busy net otherwise) Fix some warnings
git-svn-id: http://svn.inspircd.org/repository/trunk/inspircd@4450 e03df62e-2008-0410-955e-edbf42e46eb7

+++ src/inspircd.cpp
@@ -923,7 +923,7 @@ int InspIRCd::Run()
 	/* Add the listening sockets used for client inbound connections
 	 * to the socket engine
 	 */
-	for (int count = 0; count < stats->BoundPortCount; count++)
+	for (unsigned long count = 0; count < stats->BoundPortCount; count++)
 		SE->AddFd(Config->openSockfd[count],true,X_LISTEN);
 
 	this->WritePID(Config->PID);
